requirements.txt
discord.py>=2.3.0
python-dotenv>=1.0.0
aiohttp>=3.9.0
<<<<<<< HEAD
hello-world>=0.5.0
=======
httpx>=0.26.0
>>>>>>> 875cef2 (refactoring all)
